Moved bags to their own page, along with create bags. Renamed some buttons
Former-commit-id: 40a17ca8d5b8414719fe5a085a65ec459c55b837

+++ intermine/web/main/src/org/intermine/web/bag/InterMinePrimitiveBag.java
@@ -29,8 +29,8 @@ public class InterMinePrimitiveBag extends InterMineBag
      * @param size the size of the bag
      * @param os the ObjectStore to use to retrieve the contents of the bag
      */
-    public InterMinePrimitiveBag(Integer userId, String name, int size, ObjectStore os) {
-        super(userId, name, size, os);
+    public InterMinePrimitiveBag(Integer userId, String name, Integer size, ObjectStore os) {
+        super(userId, name, size.intValue(), os);
     }
 
     /**
